Refactor coroutine support namespace
