+++ src/Support/Coroutine/Drivers/SyncCoroutineDriver.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace Allnetru\Sharding\Support\Swoole;
+namespace Allnetru\Sharding\Support\Coroutine\Drivers;
 
+use Allnetru\Sharding\Support\Coroutine\CoroutineChannel;
+use Allnetru\Sharding\Support\Coroutine\CoroutineDriver;
 use Closure;
 
 /**
