+++ tests/Unit/Support/Coroutine/CoroutineDispatcherConfigurationTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Allnetru\Sharding\Tests\Unit\Support\Swoole;
+namespace Allnetru\Sharding\Tests\Unit\Support\Coroutine;
 
-use Allnetru\Sharding\Support\Swoole\CoroutineDispatcher;
+use Allnetru\Sharding\Support\Coroutine\CoroutineDispatcher;
 use Allnetru\Sharding\Tests\Stubs\FakeCoroutineDriver;
 use Allnetru\Sharding\Tests\TestCase;
 
 /**
- * @covers \Allnetru\Sharding\Support\Swoole\CoroutineDispatcher
+ * @covers \Allnetru\Sharding\Support\Coroutine\CoroutineDispatcher
  */
 class CoroutineDispatcherConfigurationTest extends TestCase
 {
